propertyApi seeding data

+++ src/Web/Mvc/Models/PaginationViewModel.cs
@@ -5,11 +5,11 @@ public class PaginatedItemsViewModel<TEntity> where TEntity : class
 
       public int PageSize { get; private set; }
 
-      public long Count { get; private set; }
+      public double Count { get; private set; }
 
       public IEnumerable<TEntity> Data { get; private set; }
 
-      public PaginatedItemsViewModel(int pageIndex, int pageSize, long count, IEnumerable<TEntity> data)
+      public PaginatedItemsViewModel(int pageIndex, int pageSize, double count, IEnumerable<TEntity> data)
       {
             PageIndex = pageIndex;
             PageSize = pageSize;
